requirements.txt
# Core
streamlit>=1.39.0
sentence-transformers>=3.0.0
numpy>=1.24.0

# PyTorch CPU-only (keeps deployment under size limits)
--extra-index-url https://download.pytorch.org/whl/cpu
torch

# Data & HTTP
httpx>=0.27.0
pandas>=2.2.0

# Export
openpyxl>=3.1.0

# File extraction
PyPDF2>=3.0.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
